+++ src/QRest.Compiler.Standard/Assembler/StandardAssembler.Context.cs
@@ -1,13 +1,15 @@
 ﻿using QRest.Core;
+using QRest.Core.Contracts;
 using QRest.Core.Expressions;
+using System;
 using System.Collections.Generic;
 using System.Dynamic;
 using System.Linq;
 using System.Linq.Expressions;
 
-namespace QRest.Compiler.Standard
+namespace QRest.Compiler.Standard.Assembler
 {
-    public partial class StandardAssembler
+    public partial class StandardAssembler : IAssemblerContext
     {
         public IReadOnlyDictionary<string, Expression> GetInitializers(IReadOnlyList<Expression> arguments)
         {
@@ -97,6 +99,41 @@ namespace QRest.Compiler.Standard
             } while (initializers.ContainsKey(name));
 
             return name;
+        }
+
+        public virtual (Expression Left, Expression Right) Convert(Expression left, Expression right)
+        {
+            var leftType = /*left.NodeType == ExpressionType.Lambda ? ((LambdaExpression)left).ReturnType : */left.Type;
+            var rightType = /*right.NodeType == ExpressionType.Lambda ? ((LambdaExpression)right).ReturnType :*/ right.Type;
+
+            if (TryConvert(right, leftType, out var newright))
+                return (left, newright);
+            else if (TryConvert(left, rightType, out var newleft))
+                return (newleft, right);
+            else
+                throw new ExpressionCreationException($"Cannot cast {leftType.Name} and {rightType.Name} either way.");
+        }
+
+        public virtual bool TryConvert(Expression expression, Type target, out Expression result)
+        {
+            if (expression.Type == target)
+            {
+                result = expression;
+                return true;
+            }
+            else if (target.IsAssignableFrom(expression.Type))
+            {
+                result = Expression.Convert(expression, target);
+                return true;
+            }
+            else if (expression.Type == typeof(string))
+            {
+                result = _options.StringParsing.Parse(expression, target);
+                return result != null;
+            }
+
+            result = null;
+            return false;
         }
     }
 }
