+++ src/QRest.Compiler.Standard/Assembler/StandardAssembler.cs
@@ -1,6 +1,4 @@
-﻿using QRest.Core;
-using QRest.Core.Contracts;
-using QRest.Core.Expressions;
+﻿using QRest.Core.Expressions;
 using QRest.Core.Extensions;
 using QRest.Core.Terms;
 using System;
@@ -8,13 +6,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 
-namespace QRest.Compiler.Standard
+namespace QRest.Compiler.Standard.Assembler
 {
-    public partial class StandardAssembler : TermVisitor, IAssemblerContext
+    public partial class StandardAssembler : TermVisitor
     {
-        private readonly StandardCompillerOptions _options;
+        private readonly IAssemblerOptions _options;
 
-        public StandardAssembler(StandardCompillerOptions options)
+        public StandardAssembler(IAssemblerOptions options)
         {
             _options = options;
         }
@@ -141,41 +139,6 @@ namespace QRest.Compiler.Standard
             , Expression.Call(typeof(Enumerable), nameof(Enumerable.ToArray)
             , new[] { element }, exp)), name);
             return exp;
-        }
-
-        public virtual (Expression Left, Expression Right) Convert(Expression left, Expression right)
-        {
-            var leftType = /*left.NodeType == ExpressionType.Lambda ? ((LambdaExpression)left).ReturnType : */left.Type;
-            var rightType = /*right.NodeType == ExpressionType.Lambda ? ((LambdaExpression)right).ReturnType :*/ right.Type;
-
-            if (TryConvert(right, leftType, out var newright))
-                return (left, newright);
-            else if (TryConvert(left, rightType, out var newleft))
-                return (newleft, right);
-            else
-                throw new ExpressionCreationException($"Cannot cast {leftType.Name} and {rightType.Name} either way.");
-        }
-
-        public virtual bool TryConvert(Expression expression, Type target, out Expression result)
-        {
-            if (expression.Type == target)
-            {
-                result = expression;
-                return true;
-            }
-            else if (target.IsAssignableFrom(expression.Type))
-            {
-                result = Expression.Convert(expression, target);
-                return true;
-            }
-            else if (expression.Type == typeof(string))
-            {
-                result = _options.StringParsing.Parse(expression, target);
-                return result != null;
-            }
-
-            result = null;
-            return false;
         }
     }
 }
