+++ src/QRest.Compiler.Standard/StringParsing/IStringParsingBehavior.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Linq.Expressions;
 
-namespace QRest.Compiler.Standard
+namespace QRest.Compiler.Standard.StringParsing
 {
     public interface IStringParsingBehavior
     {
