+++ src/QRest.Compiler.Standard/StringParsing/ParseStringsToClrTypes.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Linq.Expressions;
 
-namespace QRest.Compiler.Standard
+namespace QRest.Compiler.Standard.StringParsing
 {
     public class ParseStringsToClrTypes : IStringParsingBehavior
     {
